Huge changes: single-file model now possible
Abandoned the separate files because of not practical enough. The new
model is a top-level swipe view for swiping left/right between what is
still called 'PageColumn', can be seen as a chapter. Inside of a
chapter scrolling is top to bottom. At the end of a chapter (or
e.g. the title page) the next chapter/page/whatever is activated.